Refactor config records, fix namespaces, suppress warning
Split configuration records into separate files for clarity and modularity. Corrected namespaces for DTOs and enums to use proper "Modules" structure. Updated ReviewStatusResponse to use the correct enum import. Added SuppressMessage in ValidationMessages.cs to address a false positive security warning.
