--- src/Contracts/Contracts/Modules/Ratings/DTOs/ReviewStatusResponse.cs
@@ -1,5 +1,0 @@
-using MeAjudaAi.Contracts.Contracts.Modules.Ratings.Enums;
-
-namespace MeAjudaAi.Contracts.Contracts.Modules.Ratings.DTOs;
-
-public record ReviewStatusResponse(Guid Id, EReviewStatus Status);
