+++ src/Contracts/Modules/Communications/Channels/IEmailChannel.cs
