+++ src/Contracts/Modules/Communications/Channels/ISmsChannel.cs
