+++ src/Contracts/Modules/Communications/ICommunicationsModuleApi.cs
