+++ src/Contracts/Modules/Communications/Queries/CommunicationLogQuery.cs
