+++ src/Contracts/Modules/IModuleApi.cs
