+++ src/Contracts/Modules/Locations/DTOs/CreateAllowedCityRequestDto.cs
@@ -1,4 +1,4 @@
-namespace MeAjudaAi.Contracts.Contracts.Modules.Locations.DTOs;
+namespace MeAjudaAi.Contracts.Modules.Locations.DTOs;
 
 /// <summary>
 /// Request DTO para criação de cidade permitida.
