+++ src/Contracts/Modules/Locations/DTOs/ModuleCoordinatesDto.cs
