+++ src/Contracts/Modules/Locations/DTOs/UpdateAllowedCityRequestDto.cs
@@ -1,4 +1,4 @@
-namespace MeAjudaAi.Contracts.Contracts.Modules.Locations.DTOs;
+namespace MeAjudaAi.Contracts.Modules.Locations.DTOs;
 
 /// <summary>
 /// Request DTO para atualização de cidade permitida.
