+++ src/Contracts/Modules/Providers/DTOs/ModuleProviderBasicDto.cs
