+++ src/Contracts/Modules/Ratings/DTOs/CreateReviewRequest.cs
@@ -1,3 +1,3 @@
-namespace MeAjudaAi.Contracts.Contracts.Modules.Ratings.DTOs;
+namespace MeAjudaAi.Contracts.Modules.Ratings.DTOs;
 
 public record CreateReviewRequest(Guid ProviderId, int Rating, string? Comment);
