+++ src/Contracts/Modules/Ratings/DTOs/ProviderReviewResponse.cs
@@ -1,3 +1,3 @@
-namespace MeAjudaAi.Contracts.Contracts.Modules.Ratings.DTOs;
+namespace MeAjudaAi.Contracts.Modules.Ratings.DTOs;
 
 public record ProviderReviewResponse(Guid Id, int Rating, string? Comment, DateTime CreatedAt);
