+++ src/Contracts/Modules/Ratings/DTOs/ReviewStatusResponse.cs
@@ -1,0 +1,5 @@
+using MeAjudaAi.Contracts.Modules.Ratings.Enums;
+
+namespace MeAjudaAi.Contracts.Modules.Ratings.DTOs;
+
+public record ReviewStatusResponse(Guid Id, EReviewStatus Status);
