+++ src/Contracts/Modules/SearchProviders/DTOs/ModulePagedSearchResultDto.cs
