+++ src/Contracts/Modules/SearchProviders/DTOs/ModuleSearchableProviderDto.cs
