+++ src/Contracts/Modules/SearchProviders/ISearchProvidersModuleApi.cs
